role admin, role user

+++ src/main/java/com/avitam/fantasy11/model/RoleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends MongoRepository<Role, Long> {
+public interface RoleRepository extends MongoRepository<Role, Integer> {
     Role findByName(String name);
 }
